project/pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12, <3.14"
dependencies = [
    "cua-agent[all]>=0.4.31",
    "cua-computer>=0.4.5",
    "cua-computer-server>=0.1.22",
    "cua-som>=0.1.3",
    "dotenv>=0.9.9",
    "quart>=0.20.0",
    "quart-cors>=0.8.0",
]
